test: harden Bedrock image port
